app/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
psutil==5.9.8
psycopg2-binary==2.9.9